backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
faiss-cpu
sentence-transformers
numpy
pydantic
python-dotenv
